Refactored the CMake module system, removed per-module `Module.cmake` files, and cleaned up warnings across the codebase.
- Removed `Module.cmake` from all engine modules and the custom-module example; consolidated module registration into `helios_module()` calls inside each `CMakeLists.txt`
- Refactored the CMake layer into composable primitives (`Primitives.cmake`, `ClangArchiver.cmake`, etc.)
- Reworked `ModuleBuilder.cmake`, `ModuleRegistry.cmake`, `ModuleDiscovery.cmake`, `ModuleLinking.cmake`, `TargetUtils.cmake`, and `TestUtils.cmake` to share helpers and simplify the two-pass discovery/build flow
- Merged `helios_preregister_module` into `helios_register_module` and normalized visibility parsing for cached dependency lists
- Moved Clang archiver/LTO setup out of the root `CMakeLists.txt` into `ClangArchiver.cmake` and removed `BuildDirHelper.cmake`
- Improved `DependencyFinder.cmake` with umbrella-alias support, renamed internal dep macros, and unified system-include marking via `helios_mark_system_includes`
- Fixed deferred module linking by switching the encoded delimiter from `::` to `|` to avoid ambiguity with target names
- Polished `concurrentqueue.cmake` (include-root resolution, Homebrew/CPM layout handling) and reordered includes in `async_queue.hpp` for macOS/toolchain compatibility
- Updated module `CMakeLists.txt` files (`memory` gained `USES mimalloc` and `PUBLIC platform`, `profile` defaults to `OFF`, `async` reordered TBB require)
- Cleaned up compiler warnings across tests and headers
- Fixed minor code issues in `global_alloc.cpp` (aligned allocation rounding, profile hook attributes) and `future.cpp` (loop index types)
- Expanded several PCH files (notably `memory`) with additional standard-library and platform includes
- Updated documentation to reflect the new module workflow

+++ src/async/tests/task.cpp
@@ -167,17 +167,17 @@ TEST_SUITE("helios::async::Task") {
 
       auto task_a = graph.EmplaceTask([&execution_order, &order_counter]() {
         const int index = order_counter.fetch_add(1);
-        execution_order[index] = index;
+        execution_order[static_cast<size_t>(index)] = index;
       });
 
       auto task_b = graph.EmplaceTask([&execution_order, &order_counter]() {
         const int index = order_counter.fetch_add(1);
-        execution_order[index] = index;
+        execution_order[static_cast<size_t>(index)] = index;
       });
 
       auto task_c = graph.EmplaceTask([&execution_order, &order_counter]() {
         const int index = order_counter.fetch_add(1);
-        execution_order[index] = index;
+        execution_order[static_cast<size_t>(index)] = index;
       });
 
       task_a.Precede(task_b, task_c);
@@ -210,7 +210,7 @@ TEST_SUITE("helios::async::Task") {
       });
 
       std::vector<Task> dependent_tasks;
-      for (int i = 1; i < 4; ++i) {
+      for (size_t i = 1; i < 4; ++i) {
         dependent_tasks.push_back(
             graph.EmplaceTask([&execution_order, &order_counter, i]() {
               execution_order[i] = order_counter.fetch_add(1);
@@ -308,7 +308,7 @@ TEST_SUITE("helios::async::Task") {
       std::atomic<int> order_counter{0};
 
       std::vector<Task> dependency_tasks;
-      for (int i = 0; i < 3; ++i) {
+      for (size_t i = 0; i < 3; ++i) {
         dependency_tasks.push_back(
             graph.EmplaceTask([&execution_order, &order_counter, i]() {
               execution_order[i] = order_counter.fetch_add(1);
